# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClubHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs:
    get:
      operationId: getClubs
      tags: [clubs]
      summary: Get all clubs with enrollment status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All clubs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubsListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createClub
      tags: [clubs]
      summary: Create a new club
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClubBody"
      responses:
        "201":
          description: Club created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClubResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate club name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/leading:
    get:
      operationId: getLeadingClubs
      tags: [clubs]
      summary: Get clubs led by the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Led clubs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadingClubsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/{id}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get full club details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClub
      tags: [clubs]
      summary: Update club details (leaders only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClubBody"
      responses:
        "200":
          description: Club updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClub
      tags: [clubs]
      summary: Delete club (leaders only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /clubs/{id}/enroll:
    post:
      operationId: enrollInClub
      tags: [enrollment]
      summary: Enroll the current user in a club
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/{id}/unenroll:
    post:
      operationId: unenrollFromClub
      tags: [enrollment]
      summary: Unenroll the current user from a club
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unenrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Leaders cannot unenroll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
      required:
        - success
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
      required:
        - success
...
    ClubLeaderEntry:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        email:
          type: string
      required:
        - name
        - role
        - email
    ClubSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [Committee, Union, Club, Team, Other]
        initial:
          type: string
        default_day:
          type: string
        default_location:
          type: string
        chat_link:
          type: string
        profile_photo:
          type: string
        is_enrolled:
          type: boolean
        is_leader:
          type: boolean
      required:
        - id
        - name
        - description
        - type
        - initial
        - default_day
        - default_location
        - chat_link
        - profile_photo
        - is_enrolled
        - is_leader
    ClubsListResponse:
      type: object
      properties:
        success:
          type: boolean
        clubs:
          type: array
          items:
            $ref: "#/components/schemas/ClubSummary"
      required:
        - success
        - clubs
    LeadingClub:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [Committee, Union, Club, Team, Other]
        initial:
          type: string
        default_day:
          type: string
        default_location:
          type: string
        chat_link:
          type: string
        profile_photo:
          type: string
        user_role:
          type: string
        member_count:
          type: integer
        upcoming_events_count:
          type: integer
      required:
        - id
        - name
        - description
        - type
        - initial
        - default_day
        - default_location
        - chat_link
        - profile_photo
        - user_role
        - member_count
        - upcoming_events_count
    LeadingClubsResponse:
      type: object
      properties:
        success:
          type: boolean
        clubs:
          type: array
          items:
            $ref: "#/components/schemas/LeadingClub"
      required:
        - success
        - clubs
    EventSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        event_date:
          type: string
        event_time:
          type: string
        location:
          type: string
        description:
          type: string
      required:
        - id
        - title
        - event_date
        - event_time
        - location
        - description
    ClubDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [Committee, Union, Club, Team, Other]
        initial:
          type: string
        default_day:
          type: string
        default_location:
          type: string
        chat_link:
          type: string
        profile_photo:
          type: string
        is_enrolled:
          type: boolean
        is_leader:
          type: boolean
        leaders:
          type: array
          items:
            $ref: "#/components/schemas/ClubLeaderEntry"
        upcoming_events:
          type: array
          items:
            $ref: "#/components/schemas/EventSummary"
      required:
        - id
        - name
        - description
        - type
        - initial
        - default_day
        - default_location
        - chat_link
        - profile_photo
        - is_enrolled
        - is_leader
        - leaders
        - upcoming_events
    ClubDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        club:
          $ref: "#/components/schemas/ClubDetail"
      required:
        - success
        - club
    CreateClubBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [Committee, Union, Club, Team, Other]
        default_day:
          type: string
        default_location:
          type: string
        chat_link:
          type: string
        profile_photo:
          type: string
        leaders:
          type: array
          items:
            $ref: "#/components/schemas/ClubLeaderEntry"
      required:
        - name
        - description
        - type
        - default_day
        - default_location
        - leaders
    CreateClubResponse:
      type: object
      properties:
        success:
          type: boolean
        club_id:
          type: integer
      required:
        - success
        - club_id
...
    EnrollResponse:
      type: object
      properties:
        success:
          type: boolean
        enrollment_id:
          type: integer
      required:
        - success
        - enrollment_id
...