# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClubHub API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user profile
 */
export const GetCurrentUserResponse = zod.object({
  success: zod.boolean(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    full_name: zod.string(),
    graduation_year: zod.number().nullish(),
  }),
});

/**
 * @summary Get user settings
 */
export const GetUserSettingsResponse = zod.object({
  success: zod.boolean(),
  settings: zod.object({
    full_name: zod.string(),
    email: zod.string(),
    notifications_email: zod.boolean(),
    notifications_reminders: zod.boolean(),
    notifications_new_clubs: zod.boolean(),
    notifications_chat: zod.boolean(),
    notifications_digest: zod.boolean(),
    notifications_push_mobile: zod.boolean(),
  }),
});

/**
 * @summary Update user settings
 */
export const UpdateUserSettingsBody = zod.object({
  full_name: zod.string().optional(),
  notifications_email: zod.boolean().optional(),
  notifications_reminders: zod.boolean().optional(),
  notifications_new_clubs: zod.boolean().optional(),
  notifications_chat: zod.boolean().optional(),
  notifications_digest: zod.boolean().optional(),
  notifications_push_mobile: zod.boolean().optional(),
});

export const UpdateUserSettingsResponse = zod.object({
  success: zod.boolean(),
  settings: zod.object({
    full_name: zod.string(),
    email: zod.string(),
    notifications_email: zod.boolean(),
    notifications_reminders: zod.boolean(),
    notifications_new_clubs: zod.boolean(),
    notifications_chat: zod.boolean(),
    notifications_digest: zod.boolean(),
    notifications_push_mobile: zod.boolean(),
  }),
});

/**
 * @summary Get all clubs with enrollment status
 */
export const GetClubsResponse = zod.object({
  success: zod.boolean(),
  clubs: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string(),
      type: zod.enum(["Committee", "Union", "Club", "Team", "Other"]),
      initial: zod.string(),
      default_day: zod.string(),
      default_location: zod.string(),
      chat_link: zod.string(),
      profile_photo: zod.string(),
      is_enrolled: zod.boolean(),
      is_leader: zod.boolean(),
    }),
  ),
});

/**
 * @summary Create a new club
 */
export const CreateClubBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  type: zod.enum(["Committee", "Union", "Club", "Team", "Other"]),
  default_day: zod.string(),
  default_location: zod.string(),
  chat_link: zod.string().optional(),
  profile_photo: zod.string().optional(),
  leaders: zod.array(
    zod.object({
      name: zod.string(),
      role: zod.string(),
      email: zod.string(),
    }),
  ),
});

/**
 * @summary Get clubs led by the current user
 */
export const GetLeadingClubsResponse = zod.object({
  success: zod.boolean(),
  clubs: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string(),
      type: zod.enum(["Committee", "Union", "Club", "Team", "Other"]),
      initial: zod.string(),
      default_day: zod.string(),
      default_location: zod.string(),
      chat_link: zod.string(),
      profile_photo: zod.string(),
      user_role: zod.string(),
      member_count: zod.number(),
      upcoming_events_count: zod.number(),
    }),
  ),
});

/**
 * @summary Get full club details
 */
export const GetClubParams = zod.object({
  id: zod.coerce.number(),
});

export const GetClubResponse = zod.object({
  success: zod.boolean(),
  club: zod.object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string(),
    type: zod.enum(["Committee", "Union", "Club", "Team", "Other"]),
    initial: zod.string(),
    default_day: zod.string(),
    default_location: zod.string(),
    chat_link: zod.string(),
    profile_photo: zod.string(),
    is_enrolled: zod.boolean(),
    is_leader: zod.boolean(),
    leaders: zod.array(
      zod.object({
        name: zod.string(),
        role: zod.string(),
        email: zod.string(),
      }),
    ),
    upcoming_events: zod.array(
      zod.object({
        id: zod.number(),
        title: zod.string(),
        event_date: zod.string(),
        event_time: zod.string(),
        location: zod.string(),
        description: zod.string(),
      }),
    ),
  }),
});

/**
 * @summary Update club details (leaders only)
 */
export const UpdateClubParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateClubBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  type: zod.enum(["Committee", "Union", "Club", "Team", "Other"]),
  default_day: zod.string(),
  default_location: zod.string(),
  chat_link: zod.string().optional(),
  profile_photo: zod.string().optional(),
  leaders: zod.array(
    zod.object({
      name: zod.string(),
      role: zod.string(),
      email: zod.string(),
    }),
  ),
});

export const UpdateClubResponse = zod.object({
  success: zod.literal(true),
});

/**
 * @summary Delete club (leaders only)
 */
export const DeleteClubParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteClubResponse = zod.object({
  success: zod.literal(true),
});

/**
 * @summary Get upcoming events for a club
 */
export const GetClubEventsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetClubEventsResponse = zod.object({
  success: zod.boolean(),
  events: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      event_date: zod.string(),
      event_time: zod.string(),
      location: zod.string(),
      description: zod.string(),
    }),
  ),
});

/**
 * @summary Create an event for a club (leaders only)
 */
export const CreateClubEventParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateClubEventBody = zod.object({
  title: zod.string(),
  event_date: zod.string(),
  event_time: zod.string(),
  location: zod.string(),
  description: zod.string().optional(),
});

/**
 * @summary Delete an event (leaders only)
 */
export const DeleteEventParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteEventResponse = zod.object({
  success: zod.literal(true),
});

/**
 * @summary Enroll the current user in a club
 */
export const EnrollInClubParams = zod.object({
  id: zod.coerce.number(),
});

export const EnrollInClubResponse = zod.object({
  success: zod.boolean(),
  enrollment_id: zod.number(),
});

/**
 * @summary Unenroll the current user from a club
 */
export const UnenrollFromClubParams = zod.object({
  id: zod.coerce.number(),
});

export const UnenrollFromClubResponse = zod.object({
  success: zod.literal(true),
});

/**
 * @summary Get all events for a month with enrollment status
 */
export const GetCalendarEventsQueryParams = zod.object({
  year: zod.coerce.number().optional(),
  month: zod.coerce.number().optional(),
});

export const GetCalendarEventsResponse = zod.object({
  success: zod.boolean(),
  events: zod.array(
    zod.object({
      id: zod.number(),
      club_id: zod.number(),
      club_name: zod.string(),
      title: zod.string(),
      event_date: zod.string(),
      event_time: zod.string(),
      location: zod.string(),
      description: zod.string(),
      is_enrolled: zod.boolean(),
    }),
  ),
});

/**
 * @summary Handle Google OAuth callback
 */
export const GoogleCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  error: zod.coerce.string().optional(),
});

export const GoogleCallbackResponse = zod.object({
  success: zod.boolean(),
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    full_name: zod.string(),
    graduation_year: zod.number().nullish(),
  }),
});